+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added new API to perform a camera rendering
 - Add suport for hair master node (Double kajiya kay - Lambert)
 - Added Reset behaviour in DebugMenu (ingame mapping is right joystick + B)
+- Added default HD scene at new scene creation while in HDRP.
+- Added Wizard helping to configure HDRP project.
+- Added new UI for decal material to allow remapping and scaling of some properties
+- Added cascade shadow visualisation toggle in HD shadow settings
 
 ### Fixed
 - Fixed logic to disable FPTL with stereo rendering
@@ -30,6 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - ColorPyramid compute shader passes is swapped to pixel shader passes on platforms where the later is faster (Nintendo Switch).
 - Removing the simple lightloop used by the simple lit shader
 - Whole refactor of reflection system: Workflow and performance improvement.
+- Separated Passthrough from other RenderingPath
 
 ## [5.2.0-preview] - 2018-11-27